+++ ChangeLog.md
@@ -12,6 +12,8 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ### Changed
 
+- Reduced delay in waitForReady()
+
 ### Added
 
 ### Removed